pyproject.toml
[project]
name = "optioignons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "gymnasium>=1.1.1",
    "ipykernel>=6.29.5",
    "stable-baselines3[extra]>=2.6.0",
    "torch>=2.6.0",
]
